main app/backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
PyMuPDF>=1.21.1
reportlab>=3.6.14
requests>=2.31
corsheaders>=3.14
pytest>=7.0
pytest-django>=4.5
